requirements.txt
PySide6==6.10.1
readmdict>=0.1.1
opencc-python-reimplemented>=0.1.7
requests>=2.32.5
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.10
